Implemented update API call and made significant database updates.
Added MediaResouce and MetadataInteger tables for storing urls to
media files for download, and metadata including current revision
number.  Added revision tracking column to all tables.

Update API call takes a revision number as a parameter, and returns a
list of all exhibits, sections, and resources that need to be updated,
along with the most recent revision number so that client can save it
and knows it is up to date.